+++ 9_async_work/addon.js
@@ -11,11 +11,14 @@ function printResult(type, pi, ms) {
 
 function runSync () {
   var start = Date.now();
+  // Estimate() will execute in the current thread,
+  // the next line won't return until it is finished
 	var result = addon.calculateSync(calculations);
   printResult('Sync', result, Date.now() - start)
 }
 
 function runAsync () {
+  // how many batches should we split the work in to?
 	var batches = process.argv[3] || 16;
 	var ended = 0;
 	var total = 0;
@@ -24,11 +27,14 @@ function runAsync () {
 	function done (err, result) {
 		total += result;
 
+    // have all the batches finished executing?
 		if (++ended == batches) {
 			printResult('Async', total / batches, Date.now() - start)
 		}
 	}
 
+  // for each batch of work, request an async Estimate() for
+  // a portion of the total number of calculations
 	for (var i = 0; i < batches; i++) {
 		addon.calculateAsync(calculations / batches, done);
 	}
