+++ 9_async_work/async.h
@@ -1,0 +1,3 @@
+#include <node.h>
+
+v8::Handle<v8::Value> CalculateAsync(const v8::Arguments& args);
