+++ 9_async_work/sync.h
@@ -1,0 +1,3 @@
+#include <node.h>
+
+v8::Handle<v8::Value> CalculateSync(const v8::Arguments& args);
